Remove unnecessary for-loop that is hard to read
Using the for-loop here was a hard stretch, clearly it's not a natural
fit for visiting two distinct items, which are concatenated/returned at
the same time. Thus this was an obvious item to refactor and just use a
string format statement to return them as here, or even append to a
string using three separate statements, whatever floats you boat.