Determine if game is won or it's an advantage
The old code was using negative numbers, and I thought this wasn't as
easy to understand as the logic I have replaced it with. The rationale
is determine who has the lead, this is the person to assign the
advantage or winner designation and then determine the lead, if it's 1
it's advantage, if it's 2 or greater, it's a win. Simple!