DRY up the won_point and score updating logic
There wasn't any good reason to have this repetitive indirection, also
if the score can be updated by more than a single point at a time for a
single player, then there's no need to do this in a loop since it's in
the same func and not breaking out to consider outside things. So I
replaced that with a more succinct UpdateScore func that takes a player
and a count to update their score by.